Update README.md
++)HERE’S! WAY TO WATCH Super Bowl LIX LIVE STREAMs ON TV Channel 10 Feb 2025

The Philadelphia Eagles will face the Kansas City Chiefs in Super Bowl LIX on Sunday evening to determine this season’s National Football League champion. If you want to watch the pro football action taking place inside the Caesars Superdome, here is everything you need to know.

🌐 Watch Here 🟢==►► Super Bowl: Eagles vs. Chiefs Live Stream

🌐 Live Here 🟢==►► Super Bowl: Eagles vs. Chiefs Live Stream
This will be the second time in three years that these two NFL franchises meet on the world’s biggest stage, with Kansas City taking the crown thanks to an exciting 38-35 finish in 2023. The Chiefs won it all again last season as well, setting up a possible three-peat.

Both teams were at the top of their games during the regular season and fought their way through the playoff gauntlet to set up the matchup. Philly won the NFC East Division title with a 14-3 record while K.C. went 15-2 to win the AFC West.

Super Bowl LIX is just hours away and you can watch it online without cable this year.

The two-time defending champion Kansas City Chiefs will take on the Philadelphia Eagles in this year's edition of the big game, which kicks off at 6:30pm ET on Sunday, Feb. 9 on Fox.

If the Chiefs win, it will be the first time in the history of the NFL that a team has won three championships in a row. In fact, no two-time defending champion has ever even made it to a third straight Super Bowl, so we're already in uncharted territory. If that wasn't enough of a reason to tune in, the Chiefs and Eagles played one of the best Super Bowls of all time just two years ago.

In other words, this will probably be a pretty good football game on top of the menagerie of commercials and Kendrick Lamar's highly anticipated halftime performance.

But, if you don't pay for cable, you might be wondering how you can watch the game without shelling out for a month of an expensive streaming service. The good news is you don't have to do that at all. Here is how to watch Super Bowl LIX without taking a hammer to your piggy bank.

How to watch Super Bowl LIX without cable for free This year's game is broadcast on Fox, which means color commentary will be provided by some new guy they just hired named Tom Brady. I hear he's been to a few Super Bowls before. Anyway, in previous years when Fox has broadcast the game, it was available for free via the Fox Sports app.

That is not the case this year. Instead, the game will stream for free on the streaming service Tubi, and that should include all pregame and postgame coverage, all the same commercials people who are watching on regular TV would see, and yes, Kendrick's halftime show.

You'll need to make a Tubi account to watch the game, but fortunately, Tubi is a free streaming service by nature. Tubi will be by far the simplest way for most people who don't pay for cable to watch the game.

What time is Super Bowl LIX? The Eagles take on the Chiefs in Super Bowl LIX at 6:30 p.m. ET on Sunday, Feb. 9.

Philadelphia vs. Kansas City, Super Bowl LIX info Here’s everything you need to watch the Eagles against the Chiefs in Super Bowl LIX.

When: Sunday, Feb. 9 Where: Caesars Superdome, New Orleans, Louisiana TV channel: FOX (Watch FREE with Fubo) Live stream: Fubo Are you ready for the most exciting football matchup of the year? You better be.